# arena_server/src/public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"MatchEventDTO": {
				"description": "Match Response Event DTO",
				"properties": {
					"eventType": {
						"$ref": "#/components/schemas/MatchEventType"
					},
					"matchStartedSeed": {
						"type": "string"
					},
					"drawCard": {
						"$ref": "#/components/schemas/Card"
					},
					"cardPlayedEffect": {
						"$ref": "#/components/schemas/CardEffect"
					},
					"cardPlacedToPlayAreaCard": {
						"$ref": "#/components/schemas/Card"
					},
					"cardRemovedFromBankCard": {
						"$ref": "#/components/schemas/Card"
					},
					"cardRemovedFromBankIndex": {
						"$ref": "#/components/schemas/integer"
					},
					"turnEndedIsSuccessful": {
						"type": "boolean"
					},
					"turnEndedBonusCards": {
						"items": {
							"$ref": "#/components/schemas/Card"
						},
						"type": "array"
					},
					"turnEndedDelta": {
						"$ref": "#/components/schemas/IStateDelta"
					},
					"matchEndedScores": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					},
					"matchEndedWinnerIdx": {
						"allOf": [
							{
								"$ref": "#/components/schemas/integer"
							}
						],
						"nullable": true
					},
					"matchEndedComment": {
						"type": "string"
					},
					"matchEndedTerminated": {
						"type": "boolean"
					},
					"responseToEffectType": {
						"$ref": "#/components/schemas/CardEffectType"
					},
					"responseToEffectCard": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Card"
							}
						],
						"nullable": true
					},
					"turnStartedDelta": {
						"$ref": "#/components/schemas/IStateDelta"
					},
					"playerIndex": {
						"$ref": "#/components/schemas/integer"
					}
				},
				"required": [
					"playerIndex"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ObjectIdString": {
				"type": "string",
				"example": "30d2b1b242ea506dcc6504de",
				"format": "ObjectId",
				"description": "Stringified Object Id.",
				"pattern": "[0-9a-f]{24}|[0-9a-f]{12}"
			},
			"MatchCreationParams": {
				"description": "Match creation parameters including players, drawpile, discardpile and randomseed",
				"properties": {
					"playerids": {
						"items": {
							"$ref": "#/components/schemas/ObjectIdString"
						},
						"type": "array",
						"description": "List of players. Omit to request a 'practice' match.",
						"example": [
							"269b4f148ce2269b4f148ce2",
							"269b4f148ce2269b4f148ab3"
						],
						"minItems": 2,
						"maxItems": 2
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "Tags to add to match for later filtering"
					},
					"drawPile": {
						"items": {
							"$ref": "#/components/schemas/Card"
						},
						"type": "array",
						"description": "[Debug purposes] Draw pile initial contents of cards. When omitted default will be generated.",
						"example": [
							[
								"Oracle",
								4
							],
							[
								"Anchor",
								4
							],
							[
								"Mermaid",
								5
							]
						]
					},
					"discardPile": {
						"items": {
							"$ref": "#/components/schemas/Card"
						},
						"type": "array",
						"description": "[Debug purposes] Discard pile initial contents of cards. When omitted default will be generated.",
						"example": [
							[
								"Mermaid",
								2
							],
							[
								"Hook",
								2
							]
						]
					},
					"banks": {
						"items": {
							"items": {
								"$ref": "#/components/schemas/Card"
							},
							"type": "array"
						},
						"type": "array",
						"description": "[Debug purposes] Banks initial contents of cards. When omitted match will start with empty banks."
					},
					"randomSeed": {
						"type": "string",
						"description": "[Debug purposes] Optional seed value for pseudorandom generator used for randomization of starting player, card drawing.\r\nMatch creation returns a the input or generated seed so you can recreate any game during development.\r\nTo ignore randomization and pick first item always use \"norandom\".",
						"example": "20221105HelloWorld"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"MatchDTO": {
				"description": "Match Response item",
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/MatchId"
					},
					"playerids": {
						"items": {
							"$ref": "#/components/schemas/PlayerId"
						},
						"type": "array"
					},
					"startedAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdByPlayerId": {
						"allOf": [
							{
								"$ref": "#/components/schemas/PlayerId"
							}
						],
						"nullable": true
					},
					"creationParams": {
						"$ref": "#/components/schemas/MatchCreationParams"
					},
					"lastMoveAt": {
						"type": "string",
						"format": "date-time"
					},
					"moveCount": {
						"$ref": "#/components/schemas/integer"
					},
					"turnCount": {
						"$ref": "#/components/schemas/integer"
					},
					"moveCountInTurn": {
						"allOf": [
							{
								"$ref": "#/components/schemas/integer"
							}
						],
						"nullable": true
					},
					"state": {
						"$ref": "#/components/schemas/State"
					},
					"stateAtTurnStart": {
						"$ref": "#/components/schemas/State"
					},
					"currentPlayerIndex": {
						"allOf": [
							{
								"$ref": "#/components/schemas/integer"
							}
						],
						"nullable": true
					},
					"currentPlayerId": {
						"allOf": [
							{
								"$ref": "#/components/schemas/PlayerId"
							}
						],
						"nullable": true
					},
					"playernames": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"moves": {
						"items": {
							"$ref": "#/components/schemas/MoveDTO"
						},
						"type": "array"
					},
					"drawPileSize": {
						"$ref": "#/components/schemas/integer"
					},
					"discardPileSize": {
						"$ref": "#/components/schemas/integer"
					},
					"activePlayerIndex": {
						"$ref": "#/components/schemas/integer"
					}
				},
				"required": [
					"_id",
					"playerids",
					"startedAt",
					"lastMoveAt",
					"moveCount",
					"playernames",
					"moves",
					"drawPileSize",
					"discardPileSize"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BoolLikeString": {
				"type": "string",
				"enum": [
					"0",
					"1",
					"true",
					"false"
				]
			},
			"MatchCreateResponse": {
				"properties": {
					"randomSeed": {
						"type": "string"
					},
					"id": {
						"$ref": "#/components/schemas/ObjectIdString"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"description": "Match creation result"
			},
			"ErrorResponse": {
				"properties": {
					"error": {
						"type": "string"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
			"ActionErrorResponse": {
				"properties": {
					"events": {
						"items": {
							"$ref": "#/components/schemas/MatchEventDTO"
						},
						"type": "array"
					},
					"error": {
						"type": "string"
					}
				},
				"required": [
					"error"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"basic": {
				"scheme": "basic",
				"type": "http"
			},
			"digest": {
				"scheme": "digest",
				"type": "http"
			}
		}
	},
	"info": {
		"title": "SLHPC23 Arena",
		"version": "1.0.0",
		"description": "**SAP Labs CEE Hub Programming Competition 2023 Arean server**.\n\nYou can find more information about the game and the competititon rules at [github/SLH_SPC_2022](https://github.com/afarago/SLH_SPC_2022).  \nFor a test run, you can use the crash test dummy user `000000000000000000000000/dummypass`.  \n*Note: All the APIs expect and return application/json*.",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"email": "DL SLHPC23 <DL_637A3F6466D808029A65636A@global.corp.sap>"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/matches": {
			"get": {
				"operationId": "GetMatches",
				"responses": {
					"200": {
						"description": "matches",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MatchDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves the details of all Matches.\r\nPlayers will receive own matches only",
				"summary": "Retrieves all Matches",
				"tags": [
					"Game"
				],
				"security": [
					{
						"basic": []
					}
				],
				"parameters": [
					{
						"description": "optional filter parameter in the form of ISO date or 'today'",
						"in": "query",
						"name": "at",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "today"
					},
					{
						"description": "optional filter matches where player is active at",
						"in": "query",
						"name": "active",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/BoolLikeString"
						}
					},
					{
						"description": "optional filter matches with matching tag/comma separated list of tags",
						"in": "query",
						"name": "tags",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "optionally waits with timeout for any resulting match - useful for polling when the user receives invite for a new match to avoid polling",
						"in": "query",
						"name": "wait",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/BoolLikeString"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateMatch",
				"responses": {
					"200": {
						"description": "The newly created Match id",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MatchCreateResponse"
								}
							}
						}
					},
					"400": {
						"description": "Start a match with your player id twice or your player and the dummy player to start a practice match.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Creates a new Match using the supplied parameters, leave empty for a default 'practice' Match.",
				"summary": "Creates a new Match",
				"tags": [
					"Game"
				],
				"security": [
					{
						"basic": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Match creation parameters including players, drawpile, discardpile, banks and randomseed",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MatchCreationParams",
								"description": "Match creation parameters including players, drawpile, discardpile, banks and randomseed"
							},
							"examples": {
								"Example 1": {
									"value": {}
								},
								"Example 2": {
									"value": {
										"drawPile": [
											[
												"Oracle",
												4
											],
											[
												"Anchor",
												4
											],
											[
												"Mermaid",
												5
											]
										],
										"discardPile": [
											[
												"Mermaid",
												2
											],
											[
												"Hook",
												2
											]
										]
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/matches/{id}": {
			"get": {
				"operationId": "GetMatch",
				"responses": {
					"200": {
						"description": "Match details, with al details when requesting a 'practice' Match",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MatchDTO"
								}
							}
						}
					},
					"401": {
						"description": "Match is visible to participating players only.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Match does not exist.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "Authenticated user is not the current player.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"410": {
						"description": "No action possible on finished matches.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Retrieves the details of an existing match.\r\nSupply the unique match ID and receive corresponding match details.",
				"summary": "Retrieves a Match details",
				"tags": [
					"Game"
				],
				"security": [
					{
						"basic": []
					}
				],
				"parameters": [
					{
						"description": "The requested Match Id",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/ObjectIdString"
						}
					},
					{
						"description": "optionally waits with timeout until user becomes active",
						"in": "query",
						"name": "waitactive",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/BoolLikeString"
						}
					},
					{
						"description": "optionally add events associated with the match",
						"in": "query",
						"name": "showevents",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/BoolLikeString"
						}
					}
				]
			},
			"post": {
				"operationId": "ExecuteActionForMatch",
				"responses": {
					"200": {
						"description": "List of Response events executed in response for the Action",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MatchEventDTO"
									},
									"type": "array"
								}
							}
						}
					},
					"401": {
						"description": "Match is visible only to participating players.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Match does not exist.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "Authenticated user is not the current player.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"410": {
						"description": "No action possible on finished matches.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ActionErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Execute an Action for an existing Match.\r\nSupply the unique match ID and add Action details",
				"summary": "Execute an Action for a Match",
				"tags": [
					"Game"
				],
				"security": [
					{
						"basic": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/ObjectIdString"
						}
					},
					{
						"description": "optionally waits with timeout executing the action - useful for waiting for other user to finish its action to avoid polling",
						"in": "query",
						"name": "wait",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/BoolLikeString"
						}
					}
				],
				"requestBody": {
					"description": "Match action execution parameters",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserAction",
								"description": "Match action execution parameters"
							},
							"examples": {
								"Example 1": {
									"value": {
										"etype": "Draw"
									}
								},
								"Example 2": {
									"value": {
										"etype": "EndTurn"
									}
								},
								"Example 3": {
									"value": {
										"etype": "ResponseToEffect",
										"effect": {
											"effectType": "Oracle",
											"card": [
												"Mermaid",
												5
											]
										}
									}
								},
								"Example 4": {
									"value": {
										"etype": "ResponseToEffect",
										"effect": {
											"effectType": "Oracle",
											"card": null
										}
									}
								},
								"Example 5": {
									"value": {
										"etype": "ResponseToEffect",
										"effect": {
											"effectType": "Hook",
											"card": [
												"Mermaid",
												5
											]
										}
									}
								},
								"Example 6": {
									"value": {
										"etype": "ResponseToEffect",
										"effect": {
											"effectType": "Cannon",
											"card": [
												"Mermaid",
												5
											]
										}
									}
								},
								"Example 7": {
									"value": {
										"etype": "ResponseToEffect",
										"effect": {
											"effectType": "Sword",
											"card": [
												"Mermaid",
												5
											]
										}
									}
								},
								"Example 8": {
									"value": {
										"etype": "ResponseToEffect",
										"effect": {
											"effectType": "Map",
											"card": [
												"Mermaid",
												5
											]
										}
									}
								},
								"Example 9": {
									"value": {
										"etype": "ResponseToEffect",
										"autopick": true
									}
								},
								"Example 10": {
									"value": {
										"etype": "Draw",
										"autopick": true
									}
								},
								"Example 11": {
									"value": {
										"etype": "EndTurn",
										"autopick": true
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/matches/{id}/terminate": {
			"delete": {
				"operationId": "DeleteMatch",
				"responses": {
					"200": {
						"description": ""
					},
					"401": {
						"description": "Match is visible only to participating players.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Match does not exist.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "Forceful deletion of a match by Admin, on a timeout",
				"summary": "Forceful central deletion of a Match on a timeout",
				"tags": [
					"Game"
				],
				"security": [
					{
						"basic": []
					}
				],
				"parameters": [
					{
						"description": "game id",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/ObjectIdString"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"comment": {
										"type": "string"
									},
									"winnerId": {
										"$ref": "#/components/schemas/ObjectIdString"
									}
								},
								"required": [
									"winnerId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}